MB-60446 Use proper query context in UPDATE STATISTICS
In Trinity UPDATE STATISTICS uses the new random scan for sampling
purpose, which requires a separate FETCH operation after random
document keys are retrieved. The Fetch operation requires a query
context which is currently passed in as *execution.opContext when
the UPDATE STATISTICS is called internally from cost-based ADVISE.
The issue with this process is that for ADVISE the opContext is a
dummy one (from NewOpContext()) and thus the base operator is never
set to _RUNNING state, which in turn makes the IsActive() call
returning false, resulting in the FETCH not actually fetching any
documents.

To make this work properly, we now gets the underlying
*execution.Context from the *execution.opContext and use that as
query context instead. The ADVISE process is done at planning time
and thus there is no actual execution operator yet, and thus it
should be better to use *execution.Context instead of
*execution.opContext as the actual query context during ADVISE.

Note that this is not an issue in 7.2 since in 7.2 we use random
document retriever which does not require a FETCH step. Also for
user-issued UPDATE STATISTICS command the *execution.opContext
comes from the UpdateStatistics runtime operator and has the
proper runtime state set.

Change-Id: I79a32fcd241a4e76c4f3d299adcdddc10a9b412c
Reviewed-on: https://review.couchbase.org/c/query/+/204121
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Well-Formed: Restriction Checker