MB-60446 Additional changes for management of context in index advisor
This is only intended for 7.6.1 and above.

The previous change for MB-60446 introduced a new method in query_ee.Context
to get a handle for the datastore.QueryContext (which should just be a
pointer to *execution.Context). We can use the same mechanism for getting
the datastore.QueryContext handle in the rest of the index advisor code.

Also changed the index advisor code to use query_ee.Context when available
for the context, instead of just interface{} which is used currently.

Change-Id: I4b78abb277a5f1f4180d9780e7e44d57a014eae9
Reviewed-on: https://review.couchbase.org/c/query/+/204140
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>