MB-60461 Allow hash join in recursive part of recursive CTE
Currently we do not allow hash join to be used in the recursive
part of a recursive CTE (the subquery after the UNION/ALL). This
is due to the fact that when recursive CTE was implemented we
do not allow a hash join to be reopened, and since the recursive
part needs to be executed repeatedly, hash join would not work.
We've since lift the restriction on reopen of hash join, and
thus we can also remove the restriction on using a hash join
inside the recursive part of a recursive CTE. This should
improve performance of a recursive CTE, especially if the join
inside the recursive part is relatively large.

Change-Id: I38452d23db2b150765bef4bd965b9b1aecc0fc8a
Reviewed-on: https://review.couchbase.org/c/query/+/204516
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>