MB-60552 Don't attempt refund processing without...
... a valid tenant context.

Further, it is possible that another routine is updating the errors
array whilst NeedRefund() is iterating through the array.  This could
lead to nil elements and this change attempts to handle them too.

If/when we implement the TODO item of waiting for operators prior to the
refund processing, the chance of nil elements should go away but this
checking would still be valid protection.

Change-Id: I0adb53664713bc0276267efdc44385caad280924
Reviewed-on: https://review.couchbase.org/c/query/+/204579
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>