MB-60614 Avoid asynchronous removal of CBO stats when possible
The removal of CBO stats was made asynchronous in MB-43070 where
a deadlock is possible if the removal of CBO stats is called
during scope cleanup (e.g. manifest change).

This causes issues in some cases, e.g. MB-60567 where during
backup/restore of system collection we need to synchronously
removed old entries to avoid newly added entries to be
asynchronously removed.

The fix for MB-60567 was conservative since it is fixed in
Trinity and near its release. Only backup/restore was changed
in that fix.

There are a couple of more cases where we don't risk having
a deadlock and thus we can avoid asynchronous removal of
CBO stats:
 - when user-issued UPDATE STATISTICS DELETE was done
 - when deleting from system:dictionary

Change-Id: I7d03036989c20426e5b2487bdf74e8f86e9fe6ad
Reviewed-on: https://review.couchbase.org/c/query/+/205023
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>