Conditionally built instrumentation helper functions.
These are only built with the "debug" tag and are intended to help when
instrumenting code to debug issues.  The source location information is
included automatically - for cases where just the code path is needed
calls can be inserted with empty-string content to track it.

To use the functions build with the tag "debug", e.g.

    build.sh -S -tags enterprise,debug

Change-Id: I81bdf5e4c926e235d44943a58e813e6bc05469aa
Reviewed-on: https://review.couchbase.org/c/query/+/205121
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>