MB-60919 Improve resilience to data node changes
1. Include the underlying error as the cause for bulk get errors.
2. Enable KEEPALIVE for gomemcached connections.
3. Monitor pools for host IP changes.  This is specifically for the
observed case where two data nodes were stopped and when restarted their
name:IP mappings had swapped.  If we detect this, we discard the pool so
we aren't left waiting for all pooled connections to fail / don't suffer
unnecessary "not my v-bucket" errors.
4. Don't back-off when retrying on receiving EOF for an operation - this
typically indicates the connection was severed.  We retry without delay
so as to drain the pool of "bad" connections as quickly as possible.
Once drained of "free" connections a new connection may be attempted and
the logic for (3) will kick in.

Change-Id: Ibb74c93744f00e76521f5a10e6e7a2de7fb30c10
Reviewed-on: https://review.couchbase.org/c/query/+/205144
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>