MB-60683 Only consider flex index with valid predicate
When the planner considers indexes it originally only goes into
the code for index selection when we have a non-nil predicate.
This was changed for INCLUDE MISSING feature where we could
choose an index even without a valid predicate (where we use
the INCLUDE MISSING index as effectively a primary index).
A side effect of this is that we now goes into index selection
with no valid predicate, and this causes issues for flex index
selection since that assumes we have a valid (flex) predicate.

The fix is to add a check for non-nil flex predicate before
flex index selection.

Change-Id: I355c760bd23e8239605fe6a96f953069abf84b62
Reviewed-on: https://review.couchbase.org/c/query/+/205148
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/205158
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>