MB-60762 Proper detection of duplication error in MERGE statement
There is a behavioral change for a MERGE statement between 7.2.x and
7.6.0 if the MERGE statement only has a WHEN MATCHED THEN DELETE
clause.

Our MERGE statement is not fully compliant with the SQL standard.
The SQL standard specifies that semantically in a MERGE statement
the merge actions are performed after all the match/non-match
status of each row from the source/target table are already
determined, i.e. we should not see the effect of any merge action
in determination of the match/non-match status.

Our current implementation of the MERGE statement is "pipelined" for
performance reasons, i.e. as the match/non-match status of a
document is determined, we perform the merge action on the target
document before the underlying join operation is complete. As a
side effect of this, in case of the problematic query in the bug,
we perform the DELETE operation on the target document "immediately",
which then could potentially affect the underlying join operation.

In 7.2.x, since the DELETE operation happens immediately, we then
do not detect that the deleted document could potentailly join with
a different source document as well, i.e. if there is a "duplication"
in terms of matching, for the same target document, we may not
detect this duplication if the target document is deleted before
the next match is generated. In case of the problematic MERGE
statement, this is the case and thus we complete the MERGE operation
without error.

In 7.6.0, we now attempt to "batch" DML operations for parallelism,
but the side effect of that is that the DELETE does not happen
immediately, and thus the join operation underneath now have a
chance to produce the second match from a different source document,
as a result we now return an error on multiple delete on the same
target document.

To further complicate the situation, in 7.2.x, if the same MERGE
statement is run under a transaction, we then return the same
error on duplicated delete on the same target document, since in
case of an explicit transaction, the DELETE operation is not
performed immediately either, it is recorded at the delta table
and performed at transaction commit/abort time. That's why we
can detect the duplicated delete operation. A different error
is returned if the same MERGE statement is run under an implicit
transaction, in that case we get a CAS mismatch error since we
attempt to perform DELETE on the same target document multiple
times.

For 7.6.0 and forward, we now fixed the behavior of a MERGE
statement (with DELETE action) to be compliant with the SQL
standard, i.e., if a target document (to be deleted) matches
with multiple source documents, an error will be returned.
This is true regardless of whether the MERGE statement is run
inside a transaction (explicit or implicit) or not.

We already keep track of all document ids for the target
keyspace if the document is to be updated or deleted. We now
"delay" the delete action until the end of MERGE statement
processing, such that if a target document is to be deleted,
we don't actually send the delete request to KV until we've
processed all underlying join results. This way, if the
document to be deleted matches with multiple source documents,
we will be able to detect such duplications and raise errors
accordingly (previously we may not detect duplicates since
the target document is deleted immediately).

Note that this change does not affect UPDATE operations in
a MERGE statement, i.e. if the same target document is
updated multiple times, we will raise an error. This behavior
is currently consistent regardless of whether the MERGE
statement is inside a transaction or not, and this behavior
will not change.

To get this behavior, the following changes are made:

 - when tracking document keys from target for both
   UPDATE and DELETE (in a map), set the payload value for
   the map to be false for UPDATE and true for DELETE, such
   that we can differentiate UPDATE from DELETE.
 - in case of DELETE operation, only add the document key
   to the map, and do not send DELETE request to KV until
   the underlying join is done, at the end of MERGE statement
   processing.

With these changes the behavior should be consistent
regardless of whether the MERGE statement is run
standalone, inside an explicit transaction, or inside
an implicit transaction.

Change-Id: I91aedb871ceaffec48165084956a0336775e8d7a
Reviewed-on: https://review.couchbase.org/c/query/+/205534
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Well-Formed: Restriction Checker