MB-60762 Copy over META information if inside transaction
There is an issue with the previous fix where the DELETE action
for a MERGE statement is delayed till the end of the statement
processing. The current implementation only keeps the document
key for delayed processing, and if the MERGE statement is run
inside a transaction, the DELETE operation fails due to invalid
CAS value (which is not properly set).

To fix this, the map that is used to keep track of document
keys to be deleted now is a map[string]interface{}, the key
is the document key, and the value is an (empty) AnnotatedValue
which only contains a proper document key. In case the MERGE
statement is inside a transaction, the AnnotatedValue will also
inherit the META information from the original value, such
that it contains the proper CAS information as well.

The map for keep track of document keys for UPDATE remains the
same for now, as map[string]bool.

We also introduced a new bit in n1ql_feat_ctrl, when set,
will allow multiple delete on the same document key for the
target keyspace. This is to maintain previous behavior of
MERGE.

Change-Id: Id5595022efd993389d3324d3f1983ae955fc9a1a
Reviewed-on: https://review.couchbase.org/c/query/+/205847
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>