MB-60912 Avoid turning off CBO incorrectly for joins involving ExpressionTerm/SubqueryTerm
In the planner we check for whether CBO should be used in two ways,
whether CBO is on in settings, and whether UPDATE STATISTICS has
been run on the keyspace. The second check is really meant for
KeyspaceTerm only. The issue with this bug is that when we consider
joins involving an ExpressionTerm or SubqueryTerm, we incorrectly
turned off CBO by the second check and as a result, the joins are
not costed properly in all cases. For this particular bug, join
enumeration is used and we did not calculate a proper cost for
hash join, resulting in nested-loop join being used.

The fix is to apply the second check for CBO use only for
KeyspaceTerm.

Change-Id: I9747fd2d82adde031fa4ce1b99ca65a440315b61
Reviewed-on: https://review.couchbase.org/c/query/+/206219
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>