MB-61014 Keep index with WHERE clause ...
... even if it has less number of sargable keys.

The planner picks indexes during index selection with a
variaty of criteria. For covering index scan consideration,
the current code would prefer an index if it has "better"
sargable keys than another index. What it fails to consider
is the fact that the "worse" index in this case may have
an index condition that may allow it to be used as a
covering index. In this bug the covering index is discarded
before it is considered for covering purpose.

The fix is to add (the same) check for index condition
(for non-covering index consideration, which is done later)
for covering index part of the index selection as well.

Change-Id: I9d507604b717525f967203ac3076da77103fcf70
Reviewed-on: https://review.couchbase.org/c/query/+/206590
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>