MB-61036 Merge hang when using limit
53541 introduced code to wake and alert the producer (merge) when an
action terminated, but did so only when the producer was already
waiting.  The idea was that if not waiting then the producer would be
getting further items from its input and would detect the action
stopping via the existing parent-notify mechanism.

The problem is that the stop notification and parent notification could
be "some time" apart as at least the "after items" processing takes
place between them.  This leaves a hole where many more items can be
read and sending attempted, allowing for the possibility that a send
fills the exchange and waits without ever being woken.

This fix unconditionally marks the merge exchange as stopped eliminating
the hole (since any send attempts afterwards and before the parent
notification will detect the stopped exchange).

Change-Id: I6e029faafd38d60433eaa8df64bc7d2e60888411
Reviewed-on: https://review.couchbase.org/c/query/+/206823
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>