MB-60762 Delay processing of MERGE actions
Our MERGE statement is not compliant with the SQL standard. The SQL
standard specifies that semantically in a MERGE statement the join
between source and target to determine the proper action should not
be affected by any of the MERGE actions, i.e. any change due to
MERGE actions should be performed after the join has completed.

Our current implementation of the MERGE statement is "pipelined"
for performance reasons, i.e. as the match/non-match status of a
document is determined, we perform teh appropriate MERGE action
on that document immediately, which means the action could affect
the underlying join since the join is still on-going.

For the example MERGE statement in this bug with a DELETE action,
in 7.2.x since the DELETE happens immediately, we did not detect
that there are duplicated join results for the same target
document since the target document is deleted before the second
join result is generated. As a result, the statement completes
without error. In 7.6.x since the DML operations are now batched,
we did not execute the DELETE immediately, as a result a second
match is generated and we return an error.

To further complicate the situation, in 7.2.x, if the same MERGE
statement is run under a transaction, we then return the same
error on duplicated delete on the same target document, since in
case of an explicit transaction, the DELETE operation is not
performed immediately either, it is recorded at the delta table
and performed at transaction commit/abort time. That's why we
can detect the duplicated delete operation. A different error
is returned if the same MERGE statement is run under an implicit
transaction, in that case we get a CAS mismatch error since we
attempt to perform DELETE on the same target document multiple
times.

According to the SQL standard we should be returning an error in
this case. To make the MERGE statement SQL standard compliant,
we now delay the processing of the MERGE actions until the
underlying join is complete. This applies to all the MERGE actions.

The following changes are made:

 - we now keep an array of documents for each of the MERGE actions,
   i.e. there is an array for DELETE, an array for UPDATE and an
   array for INSERT, if all three actions are specified
 - for INSERT, the document to be inserted is constructed and put
   in the insert array
 - for UPDATE, the original document is put in the update array
 - for DELETE, there are 3 cases:
    -- if RETURNING clause is specified, then the original document
       is used
    -- else if we are inside a transaction, we construct an empty
       document and inherit all META information from the original
    -- else, we construct an empty document with just the document
       key
    the document is then put in the delete array
 - we use AnnotatedArray for each array, and the arrays can be
   spilled to disk when necessary
 - after the MERGE operator exhaustes all input, i.e. the underlying
   join is done, it then process each of the arrays and send the
   document to appropriate child to process the appropriate MERGE
   action.
 - a new bit is added to n1ql_feat_ctrl to restore the legacy merge
   behavior, i.e. if the new bit is set, we revert to legacy
   "pipelined" processing of the MERGE actions.
 - if a LIMIT clause is also specified, then we make sure we do not
   add more than the LIMIT number of documents into each array
 - we also stop processing input documents if LIMIT is reached, such
   that we don't need to finish the underlying join in that case.
 - consolidated the calculation of limit into a common utility funciton.

One effect of this change is that if the MERGE statement has an error,
then previously we could have partial DML actions persists, i.e.
whatever MERGE actions done before the error detection are already
done and cannot be rolled back. Now with the delayed processing of
MERGE actions, if the MERGE statement has an error, e.g. duplicated
key error, then no MERGE action will be performed.

Change-Id: Ibacc27766a555f8f69bb3396bb3f98854406ef58
Reviewed-on: https://review.couchbase.org/c/query/+/206897
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>