MB-61014 Keep index with WHERE clause ...
... even if it has less number of sargable keys.

The planner picks indexes during index selection with a
variaty of criteria. For covering index scan consideration,
the current code would prefer an index if it has "better"
sargable keys than another index. What it fails to consider
is the fact that the "worse" index in this case may have
an index condition that may allow it to be used as a
covering index. In this bug the covering index is discarded
before it is considered for covering purpose.

The fix is to add (the same) check for index condition
(for non-covering index consideration, which is done later)
for covering index part of the index selection as well.

MB-61014 Limit UNION SCAN generation

In the customer's scenario from the original CBSE-16480, there are
3 indexes defined, one provides covering and the other two do not.
When the covering index was eliminated early (before the previous
fix), the planner spend a lot of time trying to generate a UNION
SCAN from the reminaing two indexes. Due to the presence of many
NOT EQ and NOT IN predicates, we generated an OR clause internally
with 1024 subterms and we go through each subterm trying to generate
a plan, then combining all plans into a large UNION SCAN (with 32
children in the customer's case, due to duplication among subterms).
All of the subplans for the children of the UNION SCAN are similar
to each other, using intersect scan of the two remaining indexes.
We could have used just a single intersect scan with multiple
spans for each index in this case. In fact we do generate that plan
initially, then we attempt to use the subterms of the OR clause to
generate a large UNION SCAN, then we prefer the UNION SCAN over the
single index scan since they both provide the same sargable length.

We now add a couple of additional limits when trying to generate a
UNION SCAN. This is an attempt to avoid generating a large UNION
SCAN if an alternative scan is available:
  - if we already generated a scan with the entire OR clause, and
    we detect the OR clause has a large number of subterms (> 64)
    then just use the current scan and don't try to iterate through
    the subterms of the OR clause
  - if we have both a scan from the OR clause and a UNION SCAN from
    subterms of the OR clause, if we detect the UNION SCAN has a
    large number of children (> 16) then prefer the scan from OR
    clause

Change-Id: I9d507604b717525f967203ac3076da77103fcf70
Reviewed-on: https://review.couchbase.org/c/query/+/206590
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/206594
Well-Formed: Restriction Checker
Reviewed-on: https://review.couchbase.org/c/query/+/207130