MB-61239 Include index attributes for consideration during index selection
Our current index selection process takes index keys as expressions
only, i.e. no index attributes. This works well today, since we only
have INCLUDE MISSING and DESC as index attributes currently.
For INCLUDE MISSING, we only need to check that for the entire index,
not on a per-key basis, and for DESC, we sort of ignore that in
planning except for order pushdown, i.e. we still generate "normal"
index spans in terms of low/high, and (I’m guessing) the indexer
flips it (high as start, low as stop) during actual index scan for a
DESC index key. As a result, we currently don’t need the index
attributes information during index selection.

This will need to change for vector index keys, since we cannot tell
by just looking at the index key expression itself whether it is a
vector index key, it will be in the index attributes. Thus for index
sargability and span generation process we need the index attributes
in addition to the index key expression.

This change is not specific for vector indexes, rather it is more of
a infrastructual change to make it ready for vector indexes.
Specifically, instead of representing index keys as an array of
expressions, we now represent index keys as an array of
datastore.IndexKeys, which includes both the key expression and index
attributes. This is now included in indexEntry. For ease of
processing, we keep both an array of datastore.Indexkeys as well as
an array of expressions in each index entry, since there are places
where we still need the index keys as an expression array, e.g. for
index covering determination. Thus instead of trying to recreate an
array of expressions everytime we may need it, we just construct it
once as keep it in index entry, and use either the array of
datastore.IndexKeys or the array of expressions to represent the
index keys as necessary. Same is true also of sargKeys in index
entry as well.

Change-Id: Idfaa97f154a095eb292f972be912660fd9fe55c6
Reviewed-on: https://review.couchbase.org/c/query/+/207500
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>