MB-61239 No need to match additional index keys if ...
  ... the first index key is not INCLUDE MISSING and is not sargable

This is something I noticed while working on vector index and is not
directly related to index attributes, but it's a minor change so I'm
including it in this bug.

During sargability determination for an index, if the first index key
is not sargable, and the index key is not INCLUDE MISSING, then we
cannot use this index. The current code however continue with the
rest of the index keys, and the caller then determines that the index
is not usable (min == 0, missing == false). This is unnecessary work.

Add an extra condition such that if the leading index key is not
sargable, and it's not INCLUDE MISSING, bail out early.

Change-Id: Ib4b2a701fc12bdd126bab1cb29930d4fe32aa187
Reviewed-on: https://review.couchbase.org/c/query/+/207559
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>