MB-61239 Revert part of previous changes
It turns out part of the previous change for this bug is an
overkill. We do need the index attributes for sargability, however
when we actually determine the sargability of an individual key,
it is much easier to have the key still as an expression, with an
extra boolean indicator to mark what kind of key it is. E.g. we
can easily check for this.vector and use this.key (as expression)
instead of having this.key being an datastore.IndexKey, in which
case we need to constantly check for this.key.HasAttribute(IK_VECTOR)
which is not as convenient as checking for this.vector.

Revert part of previous changes such that we keep this.key for both
&sarg{} and &sargable{} as an expression.

Change-Id: I0c4f815dc7971054a71fde9d8f3430c26f4a5162
Reviewed-on: https://review.couchbase.org/c/query/+/207584
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>