MB-61266 Use original expression for exact spans determination
For index pushdown determination, we make sure the index has
"exact spans", i.e. no false positives. For this we make sure
the index spans generated are exact, in addition, we also need
to make sure there are no other predicates, for that we use
index covering as the way for determination.

The issue with this defect is that the index defintion has a
WHERE clause in the form of: field != <const>. The query has
the same predicate, however internally we transform the NE
predicate into a pair of range predicates:
 field < <const> OR field > <const>
and currently we use the transformed predicate for covering
determination, and that fails due to the behavior of
filter covers for OR clause, which consider only terms that
are common in all subterms of OR as valid filter cover. Thus
although we include the transformed index WHERE clause in
filter cover consideration the OR clause did not end up in
filter cover, and subsequently the same OR clause in the
query predicate is not considered as covered.

This is an issue for situations where the index WHERE clause
has a predicate that gets transformed into an OR clause,
the most common of which is the NE predicate. To fix this,
we now pass down the original form of the predicate as well,
and during exact span consideration, for covering purpose,
we consider the predicate "covered" if either the predicate
itself or the original form of the predicate (if available)
can be covered by the index.

It should be noted that this works if the OR clause comes
from internal transformation. If the index WHERE clause is
specified as an OR clause, then the fix would not work, and
there is no simple fix in that case unless we want to change
the behavior of filter covers for OR clause, which has been
this way since the very beginning and may be somewhat risky.

Change-Id: I7ea92e85c7b15044d457c5d2c86ce0a4aa4efa4b
Reviewed-on: https://review.couchbase.org/c/query/+/207640
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>