MB-62785 Pause new request handling ...
... when 80% of the configured memory quota (defaults to 50% of RAM when
no explicit quota set) is reached.  When system free memory drops below
10% pause active requests and initiate a GC cycle + return of memory to
the OS.  If this doesn't recover memory to more than 10% free, terminate
the request with the largest used memory (quota) figure.  When free
memory increases beyond 10%, start releasing requests in batches of up
to 5, depending on system free memory, and in ascending memory use
order.  This drip-feeds the requests back into the system in an effort
to avoid immediately returning to having less than 10% of the system
memory free.  Once drip-feeding has completed and used memory drops
below 80%, queued requests are once again accepted for processing.

If the feature control flag 0x200000 is set along with a node-quota,
then the node-quota is the limit rather than system memory and this
logic kicks in when less than 10% of the quota remains free.

Pausing of requests pauses operators when sending items down the
pipeline.  Testing (Q1 TCO) shows little/no impact on throughput.
Operators that are reading their input but not (yet) producing output
will continue but will come to a halt typically when either switching to
produce output or when the input pipe dries up as the producer has
stopped sending.  Paused operators have a new phase "paused" which
appears in the profile information.  The overall profile timings show an
element "pausedPipeline" which is the total time the request was marked
as paused, not a sum of the time different operators were in "paused"
state.  (It is possible for a profile to show a "pausedPipeline" time
yet have no operators showing as having been paused.)

The idea behind this change is to try prevent OOM conditions that result
in the node restarting.  Better to fail requests with a decent error or
make them wait than have the process die.  However whilst we attempt to
not kill off any requests, load may dictate that we have to.

Furthermore, requests may be delayed for some time (again depending on
the load) either with their pipelines paused or in the request queue and
may thus experience an increased rate out timeouts.

The entire feature may be turned off by setting feature control bit
0x400000.

Change-Id: I665e3f0f8e5a4f79ae4738f45d1b977bcc4f5980
Reviewed-on: https://review.couchbase.org/c/query/+/212324
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>