MB-61007: Revise authorization calls that are for internal actions
Internal authorization actions should not log an audit on authorization
failures. This is to prevent the flooding of audit logs.

1. In some system keyspaces, [ example: system:functions] an
authorization check is made on individual entities. This is an internal
action.

2. In system keyspaces like system:buckets, system:keyspaces and
system:scopes, CountScan is only allowed when the user has permissions
on the system catalogs.

Since this is the fundamental auth check for the entire action - do not
consider the authorization action as internal.

3. In some system keyspaces, a check is made initially to see if the
user is an Administrator. If it is, then the individual entity checks
are skipped.

This Admin check is an internal action.

4. In some serverless code paths, the Admin check is an internal
action.

Change-Id: I9141670b30c697743db5c63ce283fadc4f384102
Reviewed-on: https://review.couchbase.org/c/query/+/208152
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>