MB-61508 Avoid intersect scan with two similar partial indexes
For rule-based index selection (RBO), if we have two similar partial
indexes defined, when comparing the two indexes, in case the "sargability"
of the two indexes are the same, there is current logic to favor the one
with more consecutive sargable index keys (minKeys). However in case of
the bug, the index with shorter minKeys is not really shorter since it
also has an "equivalent" index key as index condition.

To take this into consideration, we check whether minKeys represent all
sargable index keys, and if so, also take into consideration the
equivalent index keys in index condition. We decide to fix it this way
instead of always consider equivalent index keys with minKeys to minimize
potential regressions. Same logic also applies to maxKeys as well.

Change-Id: Ia47b428bf5bdab3e35a4cbd31481515c510bc78f
Reviewed-on: https://review.couchbase.org/c/query/+/208606
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>