MB-61526 Revise fix to not always mark and free...
... array types.  This operation affect the underlying concrete type
which may be referenced downstream still at the time reopen() is called.
It is safe to recycle everything else as the underlying types are not
directly affected so shared references continue to function.

Also ensure values that will be updated after DML statement values are
copied as they may be shared references prior to the operations.

Change-Id: Ibc88f901a773778a51fae2347d4c3e5e1988c1be
Reviewed-on: https://review.couchbase.org/c/query/+/208856
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>