MB-61549 Revision - Do not combine index spans from different array predicates
Instead of the planner keeping track of the array id used for
ANY/ANY AND EVERY clause during generation of index spans, add
arrayId as a field to the expression itself (ANY, ANY AND EVERY
and EVERY), such that the planner can just take the array id
from the expression when needed. Compared with the previous
approach which relies on the same array id being used for the
same ANY/ANY AND EVERY clause during index span generation for
different array index keys (since they go through the same
processing steps), this approach is more robust.

 - add arrayId to collPredBase, and associated functions
 - add an expression traverser to walk an expression and assign
   arrayId to ANY/ANY AND EVERY/EVERY subexpressions.
 - call the traverser early in the planning process to set up
   arrayId to expressions needed for planning purposes.
 - add arrayId to builder, such that when the new expression
   traverser is called multiple times, we don't accidentally
   use the same array id for different array expressions.
 - use the arrayId from the expression when generating index
   spans for array index keys.

Note that technically the new arrayId is not required for
EVERY expression since we don't use that for index span
generation purpose. It is added for completeness and
consistency since it shares the same base with
ANY/ANY AND EVERY expression.

Change-Id: I1351450b13fd3f2ea74ce66d0da01d07ce334865
Reviewed-on: https://review.couchbase.org/c/query/+/208920
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>