MB-61549 Do not combine index spans from different array predicates
Prior to FLATTEN KEYS support we can only have a single array index
key in an index, and when we generate index spans from different
index keys we simply "cross multiply" the index spans for each
index key, since there is no inter-key dependency.

With the support for FLATTEN KEYS, it is now possible to have
multiple array index keys, and we can have a single array
predicate (e.g. an ANY clause) that can be used to generate spans
for different array index keys. This could have potential issues
with dependency since we are (still) generating index spans for
one index key at a time, and combine them later. If the query has
multiple ANY clauses, it is possible for the different index spans
to be combined incorrectly since we could have different index spans
generated using different ANY clauses and we should not cross-combine
index spans from different ANY clauses.

To fix this, we introduce the concept of array id for an index span,
which for array index keys denote which array predicate this index
span was generated from. Then at the time of combining index spans,
for array index keys we only combine index spans if the two spans
come from the same array index predicate (same array id), otherwise
make the span for array index key looks like _WHOLE_SPAN for the
second array index key.

The fix includes the following:

 - add arrayId to TermSpans
 - add handling of arrayId in functions that manipulate/generate
   TermSpans
 - when arrayId is changed/set, make a copy of the span first if
   it contains static spans
 - when generating spans, for array index key, set the arrayId
   (representing different array predicate) when necessary
 - when combining different spans for different index keys, if
   both spans have arrayId set and they are different, meaning
   they come from different array predicates, do not combine
   the index spans, instead, modify the 2nd index span to be
   a _WHOLE_SPAN
 - this does not affect index spans from non-array index keys,
   i.e. they are still combined as before
 - nested array index keys are also handled

MB-61549 Revision - Do not combine index spans from different array predicates

Instead of the planner keeping track of the array id used for
ANY/ANY AND EVERY clause during generation of index spans, add
arrayId as a field to the expression itself (ANY, ANY AND EVERY
and EVERY), such that the planner can just take the array id
from the expression when needed. Compared with the previous
approach which relies on the same array id being used for the
same ANY/ANY AND EVERY clause during index span generation for
different array index keys (since they go through the same
processing steps), this approach is more robust.

 - add arrayId to collPredBase, and associated functions
 - add an expression traverser to walk an expression and assign
   arrayId to ANY/ANY AND EVERY/EVERY subexpressions.
 - call the traverser early in the planning process to set up
   arrayId to expressions needed for planning purposes.
 - add arrayId to builder, such that when the new expression
   traverser is called multiple times, we don't accidentally
   use the same array id for different array expressions.
 - use the arrayId from the expression when generating index
   spans for array index keys.

Note that technically the new arrayId is not required for
EVERY expression since we don't use that for index span
generation purpose. It is added for completeness and
consistency since it shares the same base with
ANY/ANY AND EVERY expression.

Change-Id: I8a421907c8a776669922ea830fc9cc2604da1c84
Reviewed-on: https://review.couchbase.org/c/query/+/208939
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>